working on suggestions
