+++ jaclang/cli/cli.py
@@ -20,7 +20,6 @@ from jaclang.compiler.passes.main.pyast_load_pass import PyastBuildPass
 from jaclang.compiler.passes.main.schedules import py_code_gen_typed
 from jaclang.compiler.passes.tool.schedules import format_pass
 from jaclang.core.constructs import Architype
-from jaclang.core.importer import JacMachine
 from jaclang.plugin.builtin import dotgen
 from jaclang.plugin.feature import JacCmd as Cmd
 from jaclang.plugin.feature import JacFeature as Jac
@@ -91,9 +90,8 @@ def run(
     base, mod = os.path.split(filename)
     base = base if base else "./"
     mod = mod[:-4]
-    jac_machine = JacMachine(base)
     if filename.endswith(".jac"):
-        ret_module = jac_machine.jac_importer(
+        ret_module = jac_import(
             target=mod,
             base_path=base,
             cachable=cache,
@@ -106,7 +104,7 @@ def run(
     elif filename.endswith(".jir"):
         with open(filename, "rb") as f:
             ir = pickle.load(f)
-            ret_module = jac_machine.jac_importer(
+            ret_module = jac_import(
                 target=mod,
                 base_path=base,
                 cachable=cache,
