+++ jaclang/plugin/default.py
@@ -32,7 +32,8 @@ from jaclang.core.constructs import (
     WalkerArchitype,
     exec_context,
 )
-from jaclang.core.importer import JacMachine
+from jaclang.core.importer import ImportPathSpec, JacImporter, PythonImporter
+from jaclang.core.jac_machine import JacMachine
 from jaclang.core.utils import traverse_graph
 from jaclang.plugin.feature import JacFeature as Jac
 from jaclang.plugin.spec import T
@@ -216,18 +217,26 @@ class JacFeatureDefaults:
     ) -> tuple[types.ModuleType, ...]:
         """Core Import Process."""
         jac_machine = JacMachine(base_path=base_path)
-        result = jac_machine.jac_importer(
-            target=target,
-            base_path=base_path,
-            absorb=absorb,
-            cachable=cachable,
-            mdl_alias=mdl_alias,
-            override_name=override_name,
-            mod_bundle=mod_bundle,
-            lng=lng,
-            items=items,
+        spec = ImportPathSpec(
+            target,
+            base_path,
+            absorb,
+            cachable,
+            mdl_alias,
+            override_name,
+            mod_bundle,
+            lng,
+            items,
         )
-        return result
+        if lng == "py":
+            import_result = PythonImporter(jac_machine).run_import(spec)
+        else:
+            import_result = JacImporter(jac_machine).run_import(spec)
+        return (
+            (import_result.ret_mod,)
+            if absorb or not items
+            else tuple(import_result.ret_items)
+        )
 
     @staticmethod
     @hookimpl
