+++ jaclang/tests/fixtures/deep/one_lev.jac
@@ -1,6 +1,6 @@
 import:jac from .deeper, snd_lev as snd_lev;
-import:jac from ..deep, deeper;
-import:jac from ., deeper as mydeeper;
+# import:jac from ..deep, deeper;
+# import:jac from ., deeper as mydeeper;
 
 can olprint -> str {
     # deeper.snd_lev.slprint(); FIXME:
